% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_meetings.R
\name{list_meetings}
\alias{list_meetings}
\title{This function returns a tibble with all the zoom meetings in a account}
\usage{
list_meetings(user_id)
}
\arguments{
\item{user_id}{the email or user id of your email}
}
\value{
a tibble
}
\description{
This function returns a tibble with all the zoom meetings in a account
}
